requirements.txt
streamlit
pandas
gpxpy
python-fitparse
plotly
google-generativeai
